requirements.txt
requests>=2.11.0
tabulate>=0.8.2
colorama>=0.3.7
humanize>=0.5.1
python-binance==0.5.15